KDTree in its own module
